+++ src/main/java/Ui.java
@@ -1,0 +1,18 @@
+public class Ui {
+    public static void unknownCommand() {
+        System.out.println("Master, I have all the knowledge in the world but I do not recognise that command," +
+                " Please wish again");
+    }
+
+    public static void startGoodbye() {
+        System.out.println("Rub my lamp to summon me again");
+        System.out.println("Good bye for now master");
+        System.out.println("*-**-**-**-**-**-**-**-**-**-****-**-****-**-****-**");
+    }
+
+    public static void startGreeting() {
+        System.out.println("A very good day to you master, I'm Blue the Genie");
+        System.out.println("What do you wish for today? Your wish is my command");
+        System.out.println("****-**-****-**-****-**-****-**-****-**-****-**-****-**-****-**");
+    }
+}
